Move DonationPreferences to the welcome package
Every other preference data object lives with its feature (e.g.
gui.groups.GroupsPreferences); the donation prompt lives in gui.welcome.
