+++ jabgui/src/main/java/org/jabref/gui/preferences/JabRefGuiPreferences.java
@@ -41,6 +41,7 @@ import org.jabref.gui.newentry.NewEntryPreferences;
 import org.jabref.gui.preview.PreviewPreferences;
 import org.jabref.gui.sidepane.SidePaneType;
 import org.jabref.gui.specialfields.SpecialFieldsPreferences;
+import org.jabref.gui.welcome.DonationPreferences;
 import org.jabref.gui.theme.Theme;
 import org.jabref.logic.citationstyle.CSLStyleLoader;
 import org.jabref.logic.exporter.BibDatabaseWriter;
